Move Zendesk response fixture to higher level for use by other tests.
